cleaning up tools
